+++ src/main/java/net/silentchaos512/gear/api/traits/ITraitCondition.java
@@ -3,9 +3,15 @@ package net.silentchaos512.gear.api.traits;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ResourceLocation;
 import net.silentchaos512.gear.api.parts.PartDataList;
+import net.silentchaos512.gear.api.parts.PartType;
+import net.silentchaos512.gear.gear.material.MaterialInstance;
+
+import java.util.Collection;
 
 public interface ITraitCondition {
     ResourceLocation getId();
 
     boolean matches(ItemStack gear, PartDataList parts, ITrait trait);
+
+    boolean matches(ItemStack gear, PartType partType, Collection<MaterialInstance> materials, ITrait trait);
 }