+++ src/main/java/net/silentchaos512/gear/traits/conditions/AndTraitCondition.java
@@ -9,12 +9,15 @@ import net.minecraft.util.JSONUtils;
 import net.minecraft.util.ResourceLocation;
 import net.silentchaos512.gear.SilentGear;
 import net.silentchaos512.gear.api.parts.PartDataList;
+import net.silentchaos512.gear.api.parts.PartType;
 import net.silentchaos512.gear.api.traits.ITrait;
 import net.silentchaos512.gear.api.traits.ITraitCondition;
 import net.silentchaos512.gear.api.traits.ITraitConditionSerializer;
+import net.silentchaos512.gear.gear.material.MaterialInstance;
 import net.silentchaos512.gear.traits.TraitSerializers;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 public class AndTraitCondition implements ITraitCondition {
@@ -47,6 +50,16 @@ public class AndTraitCondition implements ITraitCondition {
     public boolean matches(ItemStack gear, PartDataList parts, ITrait trait) {
         for (ITraitCondition child : this.children) {
             if (!child.matches(gear, parts, trait)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    @Override
+    public boolean matches(ItemStack gear, PartType partType, Collection<MaterialInstance> materials, ITrait trait) {
+        for (ITraitCondition child : this.children) {
+            if (!child.matches(gear, partType, materials, trait)) {
                 return false;
             }
         }